notion-sync/requirements.txt
notion-client==2.2.1
python-dotenv==1.0.0
GitPython==3.1.40
